requirements.txt
# Added for Modal server and client
fastapi
uvicorn
requests
pillow
opencv-python-headless
mediapipe
matplotlib
numpy
pyserial
